+++ src/solver.jl
@@ -111,7 +111,7 @@ function get_default_options()
     convhull_ebd_ibs = false
     convhull_ebd_link = false
     convhull_warmstart = true
-    convhull_no_good_cuts = true
+    convhull_no_good_cuts = false
 
     presolve_track_time = true
     presolve_bt = true